Moving sponsorship_compassion for porting
