--- __unported__/sponsorship_compassion/data/payment_terms.xml
@@ -1,126 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="True">
-		<!-- BVR -->
-        <record id="payment_term_bvr" model="account.payment.term">
-            <field name="name">BVR</field>
-            <field name="note">BVR immediate</field>
-        </record>
-        <record id="payment_term_line_direct_debit" model="account.payment.term.line">
-            <field name="name">BVR immediate</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_bvr" name="payment_id"/>
-        </record>
-		
-		<!-- LSV -->
-		<record id="payment_term_lsv" model="account.payment.term">
-            <field name="name">LSV</field>
-            <field name="note">LSV immediate</field>
-        </record>
-        <record id="payment_term_line_lsv" model="account.payment.term.line">
-            <field name="name">LSV immediate debit</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_lsv" name="payment_id"/>
-        </record>
-		<record id="payment_term_lsv_25th" model="account.payment.term">
-            <field name="name">LSV 25th</field>
-            <field name="note">LSV</field>
-        </record>
-        <record id="payment_term_line_lsv_25th" model="account.payment.term.line">
-            <field name="name">LSV 25th</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_lsv_25th" name="payment_id"/>
-        </record>
-		<record id="payment_term_lsv_multi_months" model="account.payment.term">
-            <field name="name">LSV multi-months</field>
-            <field name="note">LSV</field>
-        </record>
-        <record id="payment_term_line_lsv_multi_months" model="account.payment.term.line">
-            <field name="name">LSV multi-months</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_lsv_multi_months" name="payment_id"/>
-        </record>
-		
-		<!-- Postfinance DD -->
-		<record id="payment_term_postfinance_dd" model="account.payment.term">
-            <field name="name">Postfinance Direct Debit</field>
-            <field name="note">Postfinance direct debit immediate</field>
-        </record>
-        <record id="payment_term_line_postfinance_dd" model="account.payment.term.line">
-            <field name="name">Postfinance immediate direct debit</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_postfinance_dd" name="payment_id"/>
-        </record>
-		<record id="payment_term_postfinance_dd_25th" model="account.payment.term">
-            <field name="name">Postfinance Direct Debit 25th</field>
-            <field name="note">Postfinance direct debit</field>
-        </record>
-        <record id="payment_term_line_postfinance_dd_25th" model="account.payment.term.line">
-            <field name="name">Postfinance immediate 25th</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_postfinance_dd_25th" name="payment_id"/>
-        </record>
-		<record id="payment_term_postfinance_dd_multi_months" model="account.payment.term">
-            <field name="name">Postfinance Direct Debit multi-months</field>
-            <field name="note">Postfinance direct</field>
-        </record>
-        <record id="payment_term_line_postfinance_dd_multi_months" model="account.payment.term.line">
-            <field name="name">Postfinance immediate multi-months</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_postfinance_dd_multi_months" name="payment_id"/>
-        </record>
-		
-		<!-- Ordre permanent -->
-		<record id="payment_term_permanent_order" model="account.payment.term">
-            <field name="name">Permanent Order</field>
-            <field name="note">Permanent order immediate</field>
-        </record>
-        <record id="payment_term_line_permament_order" model="account.payment.term.line">
-            <field name="name">Permanent order immediate</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_permanent_order" name="payment_id"/>
-        </record>
-        
-        <!-- Virement bancaire -->
-		<record id="payment_term_virement" model="account.payment.term">
-            <field name="name">Bank Transfer</field>
-            <field name="note">Bank Transfer Immediate</field>
-        </record>
-        <record id="payment_term_line_virement" model="account.payment.term.line">
-            <field name="name">Bank Transfer Immediate</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_virement" name="payment_id"/>
-        </record>
-        
-        <!-- Paiement web -->
-		<record id="payment_term_web" model="account.payment.term">
-            <field name="name">Web Payment</field>
-            <field name="note">Web Payment Immediate</field>
-        </record>
-        <record id="payment_term_web_virement" model="account.payment.term.line">
-            <field name="name">Web Payment Immediate</field>
-            <field name="value">balance</field>
-            <field eval="0" name="days"/>
-            <field eval="0" name="days2"/>
-            <field eval="payment_term_web" name="payment_id"/>
-        </record>
-    </data>
-</openerp>
