--- __unported__/sponsorship_compassion/view/suspension_wizard_view.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="view_suspension_wizard" model="ir.ui.view">
-            <field name="name">project.suspension.wizard.form</field>
-            <field name="model">compassion.project.suspension.wizard</field>
-            <field eval="18" name="priority"/>
-            <field name="arch" type="xml">
-                <form string="Project suspension extension" version="7.0">
-                    <p>This will extend the suspension of the project to the defined date.</p>
-                    <group>
-                        <field name="date_start"/>
-                        <field name="date_end"/>
-                    </group>
-                    <footer>
-                        <button name="perform_suspension" string="Extend suspension" type="object" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link" special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-
-        <!-- Menu Item -->
-        <act_window id="action_suspend_project"
-                name="Suspend Project"
-                res_model="compassion.project.suspension.wizard" src_model="compassion.project"
-                view_mode="form" target="new" view_type="form"
-                />
-    </data>
-</openerp>
