+++ sponsorship_compassion/__openerp__.py
