+++ sponsorship_compassion/i18n/fr_CH.po
